requirements.txt
pymongo[srv]
flask
flask-cors
python-dotenv
scikit-learn
numpy
pandas
matplotlib
seaborn
python-dotenv
wtforms
flask-wtf
argon2-cffi
flask-login
email_validator
dhooks
flask_reuploaded
werkzeug
threading
socket
flask[async]
paypalrestsdk
browser_cookie3
requests
openai

